rename navigation component to drawerAppBar
